% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachify.character.R
\name{cachify.character}
\alias{cachify.character}
\title{cachify turns a function into a caching function}
\usage{
\method{cachify}{character}(f, cacheDir, debug = FALSE)
}
\arguments{
\item{f}{name of the function to cachify}

\item{cacheDir}{A directory where the results of the function are stored.}

\item{debug}{Set to TRUE to receive info about cache retrieval vs. evaluation}
}
\value{
No return. The original function with name \code{f} is replaced by the cachified version
}
\description{
cachify turns a function into a caching function
}
