% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachify.R
\name{cachify}
\alias{cachify}
\title{cachify turns a function into a caching function}
\usage{
cachify(f, cacheDir, debug = FALSE)
}
\arguments{
\item{f}{The function to cachify (either name or function)}

\item{cacheDir}{A directory where the results of the function are stored.}

\item{debug}{Set to TRUE to receive info about cache retrieval vs. evaluation}
}
\value{
The cachified function. Cachifistic!
}
\description{
cachify turns a function into a caching function
}
\details{
ATTENTION: functions that are called for their side-effects become
         unusable without error!
}
