% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCacheDir.R
\name{getCacheDir}
\alias{getCacheDir}
\title{Get cache directory of a cachified function}
\usage{
getCacheDir(f)
}
\arguments{
\item{f}{The function to return the cache directory from}
}
\value{
The cache directory of the function
}
\description{
Get cache directory of a cachified function
}
