% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isCacheLocked.R
\name{isCacheLocked}
\alias{isCacheLocked}
\title{Is the cache of a function locked?}
\usage{
isCacheLocked(f)
}
\arguments{
\item{f}{The function to inform about its cache status}
}
\description{
Is the cache of a function locked?
}
\seealso{
\code{\link{lockCache}}
}
