% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockCache.R
\name{lockCache}
\alias{lockCache}
\title{Lock a cache; whenever the function tries to evaluate rather than read from
the cache, it will fail with an error. Useful, if you want to ensure
reproduction of former execution}
\usage{
lockCache(f)
}
\arguments{
\item{f}{The function to lock the cache from}
}
\description{
Lock a cache; whenever the function tries to evaluate rather than read from
the cache, it will fail with an error. Useful, if you want to ensure
reproduction of former execution
}
