% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceFunctionByReference.R
\name{replaceFunctionByReference}
\alias{replaceFunctionByReference}
\title{Replaces a function with a new value, taking care of 
consistent placement in the search path}
\usage{
replaceFunctionByReference(name, newFunc, startEnv = environment())
}
\arguments{
\item{name}{character; name of the function to replace}

\item{newFunc}{function; new value}

\item{startEnv}{environment; where to start looking for 'name'}
}
\value{
no return; the first instance of 'name' on the search path 
       will be replaced by newFunc.
}
\description{
Replaces a function with a new value, taking care of 
consistent placement in the search path
}
\details{
Tries to consistently take care of locked environments
         by creating a "cachified" unlocked environment one position upstream in the search path
         and assigning the new function there. \cr
         Cachified functions need special care because they don't work without their
         enclosing environment. This is dealt with by leaving a cachified new function 
         in its enclosing environment and only setting the parent to 
         the enclosing environment of the original function
}
