% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCacheDir.R
\name{setCacheDir}
\alias{setCacheDir}
\title{Set cache directory of a cachified function}
\usage{
setCacheDir(f, cacheDir)
}
\arguments{
\item{f}{The function to set the cache directory from}

\item{cacheDir}{A directory where the results of the function are stored.}
}
\description{
Set cache directory of a cachified function
}
