% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncachify.R
\name{uncachify}
\alias{uncachify}
\title{Remove caching functionality from function}
\usage{
uncachify(f)
}
\arguments{
\item{f}{The function to remove caching functionality from}
}
\value{
The uncachified function
}
\description{
Remove caching functionality from function
}
